Refactor: App.tsx 리펙토링 > 인증 초기화 로직, 라우팅설정 분리
